Type : Feature Fix & Refactoring
1. 프로그래밍 응용 과제 5차
- totalDays 반환형을 long으로 변경. int형보다 훨씬 큰값이 나올 수 있음.
- 프로그램 상 출력과 사용자의 입력을 구분하기 위해 프로그램 상 출력은 앞에 [SYS]를 붙임.
- cout 과 cin으로 입력을 받지않고, getString과 getInt로 통일.

+++ 프로그래밍 응용 과제/Chapter 5/Chapter 5 - Library/Date.cpp
@@ -22,7 +22,7 @@ Date::Date(std::string dateStr)
 	}
 }
 
-int Date::totalDays()		//1년 1월 1일부터 currentDate까지의 일 수
+long Date::totalDays()		//1년 1월 1일부터 currentDate까지의 일 수
 {
 	int months[] = { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
 	int i;
